Use static virtualenv name
The variable thing didn't work out like I wanted.
In rule ${VENV}/virtualenv the variable $@ kept expanding either to venv
or to /virtualenv, but never venv/virtualenv, and I don't know why.
Having the virtualenv name in a variable was not really necessary to
begin with, so let's not make this more complicated than it has to be.

This happened on my Arch Linux system. If I recall correctly on my NixOS
system it worked flawlessly (that's where I started this repo).